[#4] Refactor: app.cpp 파일 내용 분리
- background: Background 클래스
- sprite: Sprite 클래스
- decode_img: 이미지 관련 함수 / 변수
