+++ src/render/render.hpp
@@ -1,0 +1,8 @@
+#ifndef RENDER_H
+#define RENDER_H
+
+#include "background.hpp"
+#include "sprite.hpp"
+#include "decode_img.hpp"
+
+#endif
